+++ models/models.go
@@ -1737,29 +1737,10 @@ type InlineQueryResultsButton struct {
 }
 
 // contains subtypes
-type InlineQueryResult struct {
-	Type InlineQueryResultType
+type InlineQueryResult interface {
+	isInlineQueryResult()
 
-	InlineQueryResultCachedAudio    *InlineQueryResultCachedAudio
-	InlineQueryResultCachedDocument *InlineQueryResultCachedDocument
-	InlineQueryResultCachedGif      *InlineQueryResultCachedGif
-	InlineQueryResultCachedMpeg4Gif *InlineQueryResultCachedMpeg4Gif
-	InlineQueryResultCachedPhoto    *InlineQueryResultCachedPhoto
-	InlineQueryResultCachedSticker  *InlineQueryResultCachedSticker
-	InlineQueryResultCachedVideo    *InlineQueryResultCachedVideo
-	InlineQueryResultCachedVoice    *InlineQueryResultCachedVoice
-	InlineQueryResultArticle        *InlineQueryResultArticle
-	InlineQueryResultAudio          *InlineQueryResultAudio
-	InlineQueryResultContact        *InlineQueryResultContact
-	InlineQueryResultGame           *InlineQueryResultGame
-	InlineQueryResultDocument       *InlineQueryResultDocument
-	InlineQueryResultGif            *InlineQueryResultGif
-	InlineQueryResultLocation       *InlineQueryResultLocation
-	InlineQueryResultMpeg4Gif       *InlineQueryResultMpeg4Gif
-	InlineQueryResultPhoto          *InlineQueryResultPhoto
-	InlineQueryResultVenue          *InlineQueryResultVenue
-	InlineQueryResultVideo          *InlineQueryResultVideo
-	InlineQueryResultVoice          *InlineQueryResultVoice
+	MarshalCustom() ([]byte, error)
 }
 
 type InlineQueryResultArticle struct {